fix: pagination issues
