+++ pages/small-spender/index.tsx
@@ -2,9 +2,7 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { Pagination } from 'antd';
 import { GetServerSideProps } from 'next';
-import { useRouter } from 'next/router';
 import { getOnChainPosts, IPostsListingResponse } from 'pages/api/v1/listing/on-chain-posts';
 import { IReferendumV2PostsByStatus } from 'pages/root';
 import React, { FC, useEffect } from 'react';
@@ -20,7 +18,6 @@ import SEOHead from '~src/global/SEOHead';
 import { sortValues } from '~src/global/sortOptions';
 import { IApiResponse, PostOrigin } from '~src/types';
 import { ErrorState } from '~src/ui-components/UIStates';
-import { handlePaginationChange } from '~src/util/handlePaginationChange';
 
 export const getServerSideProps: GetServerSideProps = async ({ req, query }) => {
 	const { page = 1, sortBy = sortValues.NEWEST, filterBy } = query;
@@ -113,16 +110,6 @@ const SmallSpender: FC<ISmallSpenderProps> = (props) => {
 	// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, []);
 
-	const router = useRouter();
-	const onPaginationChange = (page:number) => {
-		router.push({
-			query:{
-				page
-			}
-		});
-		handlePaginationChange({ limit: LISTING_LIMIT, page });
-	};
-
 	if (error) return <ErrorState errorMessage={error} />;
 
 	if (!posts || Object.keys(posts).length === 0) return null;
@@ -132,16 +119,6 @@ const SmallSpender: FC<ISmallSpenderProps> = (props) => {
 			trackName={PostOrigin.SMALL_SPENDER}
 			posts={posts}
 		/>
-		<div className='flex justify-end mt-6'>
-			<Pagination defaultCurrent={1}
-				pageSize={LISTING_LIMIT}
-				total={posts.all?.data?.count || 0}
-				showSizeChanger={false}
-				hideOnSinglePage={true}
-				onChange={onPaginationChange}
-				responsive={true}
-			/>
-		</div>
 	</>;
 };
